Move oidc implementation to shell/imports/oidc
