+++ shell/imports/client/accounts/login-buttons.js
@@ -33,8 +33,8 @@ import {
 } from "/imports/client/accounts/email-token/token-login-helpers.js";
 import { GrainViewList } from '/imports/client/grain/grainview-list.js';
 import { loginWithLDAP } from "/imports/client/accounts/ldap/ldap-client.js";
-import { loginWithOidc } from "/imports/client/accounts/oidc/oidc-client.js";
 import { loginWithSaml } from "/imports/client/accounts/saml/saml-client.js";
+import { loginWithOidc } from "/imports/oidc/oidc-client.js";
 import AccountsUi from "/imports/client/accounts/accounts-ui.js";
 import { SandstormDb } from "/imports/sandstorm-db/db.js";
 import { globalDb } from "/imports/db-deprecated.js";
