+++ shell/server/main.ts
@@ -60,12 +60,13 @@ import "../imports/blackrock-payments/constants.js";
 import "../imports/blackrock-payments/server/payments-server.js";
 import "../imports/blackrock-payments/server/payments-api-server.js";
 
+// oidc
+import "../imports/oidc/oidc-server.js";
 
 // Import everything from server/ in the order that Meteor would have.
 import "../imports/server/accounts/credentials/credentials-server.js";
 import "../imports/server/accounts/email-token/token-server.js";
 import "../imports/server/accounts/ldap/ldap-server.js";
-import "../imports/server/accounts/oidc/oidc-server.js";
 import "../imports/server/accounts/saml/saml-server.js";
 import "../imports/server/accounts/accounts-server.js";
 import "../imports/server/accounts/accounts-ui-methods.js";
